metaretriever/requirements.txt
absl-py==1.0.0
altair==4.2.0
anyio==3.5.0
anytree==2.8.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
black==21.12b0
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
colorlog==6.6.0
commonmark==0.9.1
conllu==4.4.1
cycler==0.11.0
dataclasses==0.6
datasets==1.9.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
elasticsearch==7.16.3
entrypoints==0.3
executing==0.8.2
faiss-cpu==1.7.2
fastapi==0.74.1
filelock==3.0.12
fire==0.4.0
fonttools==4.28.5
fsspec==2022.1.0
future==0.18.2
git-python==1.0.3
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.3.3
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
grpcio==1.43.0
h11==0.13.0
h5py==3.6.0
huggingface-hub==0.0.8
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.7.0
ipython
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
jieba==0.42.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.1
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multiprocess==0.70.12.2
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.7
notebook==6.4.7
numpy==1.19.5
oauthlib==3.1.1
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
pluggy==1.0.0
portalocker==2.3.2
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.24
protobuf==3.19.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.1
py==1.11.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pydeck==0.7.1
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.6
pyrsistent==0.18.1
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.0
rich==9.8.2
rouge-score==0.0.4
rsa==4.8
sacrebleu==1.4.14
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
stack-data==0.1.4
starlette==0.17.1
streamlit==1.4.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.4.0
tensorflow-metadata==1.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.6.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
uvicorn==0.17.5
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
xxhash==2.0.2
zipp==3.7.0
learn2learn